+++ geode-core/src/main/java/org/apache/geode/cache/Region.java
@@ -36,7 +36,6 @@ import org.apache.geode.cache.query.QueryService;
 import org.apache.geode.cache.query.SelectResults;
 import org.apache.geode.cache.query.TypeMismatchException;
 import org.apache.geode.cache.snapshot.RegionSnapshotService;
-import org.apache.geode.internal.logging.EntriesCollection;
 
 /**
  * Manages subregions and cached data. Each region can contain multiple subregions and entries for
@@ -141,7 +140,7 @@ import org.apache.geode.internal.logging.EntriesCollection;
  * @since GemFire 2.0
  */
 
-public interface Region<K, V> extends ConcurrentMap<K, V>, EntriesCollection {
+public interface Region<K, V> extends ConcurrentMap<K, V> {
   /** The region name separator character. */
   char SEPARATOR_CHAR = '/';
