+++ geode-core/src/main/java/org/apache/geode/internal/cache/AbstractRegion.java
@@ -106,7 +106,7 @@ import org.apache.geode.pdx.internal.PeerTypeRegistration;
  */
 @SuppressWarnings("deprecation")
 public abstract class AbstractRegion implements InternalRegion, AttributesMutator, CacheStatistics,
-    DataSerializableFixedID, Extensible<Region<?, ?>>, EvictableRegion {
+    DataSerializableFixedID, Extensible<Region<?, ?>>, EvictableRegion, EntriesCollection {
 
   private static final Logger logger = LogService.getLogger();
   private final ReentrantReadWriteLock readWriteLockForCacheLoader = new ReentrantReadWriteLock();
