+++ geode-core/src/main/java/org/apache/geode/internal/cache/EntriesCollection.java
@@ -12,7 +12,12 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal.logging;
+package org.apache.geode.internal.cache;
 
+import org.apache.geode.cache.Region;
+
+/**
+ * Collection of data entries in a data structure such as a {@link Region}.
+ */
 public interface EntriesCollection {
 }
