+++ geode-core/src/main/java/org/apache/geode/internal/cache/EntriesSet.java
@@ -26,7 +26,6 @@ import org.apache.geode.cache.EntryDestroyedException;
 import org.apache.geode.cache.Region;
 import org.apache.geode.internal.cache.LocalRegion.IteratorType;
 import org.apache.geode.internal.cache.entries.AbstractRegionEntry;
-import org.apache.geode.internal.logging.EntriesCollection;
 
 /** Set view of entries */
 public class EntriesSet extends AbstractSet implements EntriesCollection {
